remove ActionElement
