+++ web_ui/src/pages/annotator/components/labels/label-tag/label-tag.component.tsx
@@ -4,8 +4,8 @@
 import { Flex, Tooltip, TooltipTrigger } from '@adobe/react-spectrum';
 
 import { Label } from '../../../../../core/labels/label.interface';
-import { PressableElement } from '../../../../../shared/components/action-element/action-element.component';
 import { LabelColorThumb } from '../../../../../shared/components/label-color-thumb/label-color-thumb.component';
+import { PressableElement } from '../../../../../shared/components/pressable-element/pressable-element.component';
 import { TruncatedText } from '../../../../../shared/components/truncated-text/truncated-text.component';
 
 interface LabelTagProps {
