+++ web_ui/src/pages/annotator/components/sidebar/dataset/dataset-list-item-details.component.tsx
@@ -10,8 +10,8 @@ import { usePress } from 'react-aria';
 import { isVideo } from '../../../../../core/media/video.interface';
 import { useUsers } from '../../../../../core/users/hook/use-users.hook';
 import { useOrganizationIdentifier } from '../../../../../hooks/use-organization-identifier/use-organization-identifier.hook';
-import { PressableElement } from '../../../../../shared/components/action-element/action-element.component';
 import { MediaItemView } from '../../../../../shared/components/media-item-view/media-item-view.component';
+import { PressableElement } from '../../../../../shared/components/pressable-element/pressable-element.component';
 import { TruncatedTextWithTooltip } from '../../../../../shared/components/truncated-text/truncated-text.component';
 import { getFileSize } from '../../../../../shared/utils';
 import { getMediaId } from '../../../../media/utils';
@@ -78,9 +78,7 @@ export const DatasetListItemDetails = ({
                     </View>
                     <Flex direction={'column'} width={'100%'}>
                         <TooltipTrigger placement={'bottom'}>
-                            <PressableElement UNSAFE_className={classes.itemMediaName}>
-                                <Text>{name}</Text>
-                            </PressableElement>
+                            <PressableElement UNSAFE_className={classes.itemMediaName}>{name}</PressableElement>
                             <Tooltip>{name}</Tooltip>
                         </TooltipTrigger>
