+++ web_ui/src/pages/annotator/components/sidebar/toggle-sidebar-button.component.tsx
@@ -10,7 +10,7 @@ import ChevronDoubleLeft from '@spectrum-icons/workflow/ChevronDoubleLeft';
 import ChevronDoubleRight from '@spectrum-icons/workflow/ChevronDoubleRight';
 import { useToggleButton } from 'react-aria';
 
-import { PressableElement } from '../../../../shared/components/action-element/action-element.component';
+import { PressableElement } from '../../../../shared/components/pressable-element/pressable-element.component';
 
 import classes from './toggle-sidebar.module.scss';
