+++ web_ui/src/pages/project-details/components/project-media/grid-media-item.component.tsx
@@ -7,9 +7,9 @@ import { Flex, Tooltip, TooltipTrigger } from '@adobe/react-spectrum';
 import { View } from '@react-spectrum/view';
 
 import { MediaItem } from '../../../../core/media/media.interface';
-import { PressableElement } from '../../../../shared/components/action-element/action-element.component';
 import { Checkbox } from '../../../../shared/components/checkbox/checkbox.component';
 import { MediaItemView } from '../../../../shared/components/media-item-view/media-item-view.component';
+import { PressableElement } from '../../../../shared/components/pressable-element/pressable-element.component';
 import { MediaItemActions } from './media-item-actions/media-item-actions.component';
 import { MediaItemTooltipMessage } from './media-item-tooltip-message/media-item-tooltip-message';
 import { getMediaItemTooltipProps } from './media-item-tooltip-message/utils';
