+++ web_ui/src/shared/components/header/jobs-management/jobs-list-item-status.component.tsx
@@ -12,8 +12,8 @@ import { JobStepState } from '../../../../core/jobs/jobs.const';
 import { Job, JobStep } from '../../../../core/jobs/jobs.interface';
 import { ANIMATION_PARAMETERS } from '../../../animation-parameters/animation-parameters';
 import { isNonEmptyArray } from '../../../utils';
-import { PressableElement } from '../../action-element/action-element.component';
 import { LoadingIndicator } from '../../loading/loading-indicator.component';
+import { PressableElement } from '../../pressable-element/pressable-element.component';
 import { QuietActionButton } from '../../quiet-button/quiet-action-button.component';
 import { ThinProgressBar } from '../../thin-progress-bar/thin-progress-bar.component';
 import { JobsListItemDetailedProgress } from './jobs-list-item-progress.component';
