+++ web_ui/src/shared/components/label-tree-view/label-tree-view-item/group-edition-mode/group-creation.component.tsx
@@ -17,9 +17,9 @@ import { LabelsRelationType } from '../../../../../core/labels/label.interface';
 import { filterGroups } from '../../../../../core/labels/utils';
 import { newLabelNameSchema } from '../../../../../pages/create-project/components/utils';
 import { isYupValidationError } from '../../../../../pages/user-management/profile-page/utils';
-import { PressableElement } from '../../../action-element/action-element.component';
 import { Button } from '../../../button/button.component';
 import { LimitedTextField } from '../../../limited-text-field/limited-text-field.component';
+import { PressableElement } from '../../../pressable-element/pressable-element.component';
 import { ValidationErrorMsg } from '../../../validation-error-msg/validation-error-msg.component';
 import { getNewGroup } from '../../utils';
 import { MULTIPLE_SELECTION_SWITCH_TOOLTIP, NEW_GROUP_FIELD_TOOLTIP } from './utils';
