+++ web_ui/src/shared/components/label-tree-view/label-tree-view-item/group-edition-mode/group-edition-mode.component.tsx
@@ -13,8 +13,8 @@ import { LabelsRelationType } from '../../../../../core/labels/label.interface';
 import { newLabelNameSchema } from '../../../../../pages/create-project/components/utils';
 import { isYupValidationError } from '../../../../../pages/user-management/profile-page/utils';
 import { idMatchingFormat } from '../../../../../test-utils/id-utils';
-import { PressableElement } from '../../../action-element/action-element.component';
 import { LimitedTextField } from '../../../limited-text-field/limited-text-field.component';
+import { PressableElement } from '../../../pressable-element/pressable-element.component';
 import { ValidationErrorMsg } from '../../../validation-error-msg/validation-error-msg.component';
 import { getEditedItem } from '../../utils';
 import { ItemEditionState } from '../item-edition-state/item-edition-state.component';
